requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cloudinary==1.25.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
gunicorn==20.1.0
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.2
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
